Enable saving settings capability on USB or HDD.
As requirement, you have to create an (empty or copy an existing one) conf_opl.cfg on the targeted device (you can also copy other configuration files: DNAS, VMC, compat).

So the complete process will be (as Dlanor described it, with some precision):
* OPL start and try to load the conf_opl.cfg from any MC
* If no conf file is found on MCs, then it will try to load the conf_opl.cfg on mass (all partitions)
* If no conf file is found on mass partitions, then it will try load the conf_opl.cfg from hdd

If configuration files are found on one of the USB/HDD device, OPL will use them for saving for the whole session. Meaning that if you plug later an MC with configuration files on it, OPL will not swap and use them.

Known limitation: you may expect the conf_apps.cfg to be loaded from the alternate device ... but it is not ... for the moment. Will do it later probably.