* IPCONFIG.DAT file now loaded from the alternate device too * small correction for the alternate device mode
When no config for OPL is found:
> if there is any MCs connected, then OPL will continue directly in "first start" mode
> if no MCs inserted at all, OPL will scan for config on alternate devices until success

If you want to start OPL in "alternate device" mode and have no config file on it, then OPL will stay on black screen, constantly polling the device (as described above). In that case press and hold "circle" (or "cross"), then OPL will abort attempts and continue without any config set, but the subsequent config saves will happen on mass (or HDD if no mass present).