* Minor GUI changes (deferred error message display, small optim for main loop) * provide Error messages for network startup
On error when starting the network, you will have a message with an error code, which correspond to

#define ERROR_ETH_NOT_STARTED			100
#define ERROR_ETH_MODULE_PS2DEV9_FAILURE	200
#define ERROR_ETH_MODULE_SMSUTILS_FAILURE	201
#define ERROR_ETH_MODULE_SMSTCPIP_FAILURE	202
#define ERROR_ETH_MODULE_SMSMAP_FAILURE		203
#define ERROR_ETH_MODULE_SMBMAN_FAILURE		204
#define ERROR_ETH_SMB_LOGON			300
#define ERROR_ETH_SMB_ECHO			301
#define ERROR_ETH_SMB_OPENSHARE			302

If you have an error 300, it means OPL fails connecting to your SAMBA/CIFS serveur, so it can be due to:
* bad IP address for your PC
* bad port
* false crendential (or guest account not allowed)

If you have an error 302 is when the share name is invalid.

Finally, YOU CAN RE-TRY NOW TO CONNECT AS MANY TIMES AS NEEDED (until successful connection).

To do so, try connecting, check the error code, go to network settings, change them, go back to your menu list and HIT REFRESH ("select" button) OR WAIT FOR AUTO-REFRESH IF ENABLED. OPL will try to connect again.