Added a proper patch for Choujikuu Yousai Macross (SLPM-65405).
As commented within the source code:
It appears that Choujikuu Yosai Macross has a rather large problem with it: it appears to be using its GUI before initialization is completed.

I did not figure out whether the crash at startup is caused by a real timing problem (whereby the GUI is used before initialization is completed by another thread, if there was one), or if its normal functionality was the result of pure luck.

By right, this fix should be applied to ALL modes, and not just HDD mode. But I think that it's better to patch the game only when required, at least until this patch has been proven to have no side effects.