Updated to support the new PS2SDK revision, changed module management so that only the necessary modules will remain resident in-game, changed unloadPad() to call padEnd() instead of padReset() in order to actually stop the PADMAN RPC service, re-added the "ei" instruction to crt0, fixed the logic that determines whether the IP address config has been changed (When DHCP is enabled, re-arranged the EE core so that modules (0x0009A000) will now be stored after the EE core (0x00088000).
Regarding the re-addition of the "ei" instruction to crt0: while it was removed because it isn't actually necessary because ExecPS2() will start the EE core with interrupts enabled, the kernel's LoadExecPS2() implementation to use will cause it to be started with interrupts disabled and in kernel mode.
It should remain there, so that it won't have to be re-added if the EE core is shifted into kernel RAM or used with the kernel LoadExecPS2() syscall implementation.