Revert "Optional FMV fix (enabled by setting UNALIGNED_BUFFER_PATCH to 0)" Revert "I also like to commit in two-phases...:p"
Not only will there be no additional benefits, it will fail if the buffer is not aligned.
The original code already allows for performance when the buffer is aligned.

The original code was there to accommodate the restrictions imposed by the USBD module.
Due to hardware errata, buffers must be properly aligned. Or 63/64-byte frames be limited to 62, but not all devices are compatible with this other method.

The hardware errata seems to only concern the bulk-out pipe and not bulk-in.
However, the USBD module performs the alignment check on the pipe, as long as it was opened with sceUsbdOpenPipeAligned().

This reverts commit b6418d20786a4a806d3600dd9a7b19f74d9ea87f.
This reverts commit 68112d7f65236b8e66d7dcd18877e2e5cc9acebc.