(mass_driver.c) Fix glitches related to USB support that were introduced with commit 18780d5 on June 8th.
1. Some devices to cease being compatible due to SET INTERFACE failing.
By the USB 1.1 specification, section 9.4.10: Devices that only support a default setting for the specified interface may return a STALL.
If that happens, we shall clear the halt state of the interface's pipes and continue, as with Linux.

2. Fixed faulty logic for reading data from devices with sector sizes > 2048 bytes.
As the code for making reads in blocks was removed, the loop only iterates once.
That is wrong, as reading from a 4096-byte device (for example) could require the loop to iterate twice if the last LBA is not aligned with the device's physical LBAs.
At the 2nd iteration, nsectors would have been 1, when it should then read a complete sector and copy the first half of it with memcpy.