Added workaround for Ridge Racer V (SLUS-20002 & SLES-50000), for USB mode. The game seems to have problems when the media takes too long to completely read data. The data processed by the game appears to get offset, causing some loop to never end, which ends with the scratchpad overflowing.
Unfortunately, I could not understand how the whole reading system worked.
This workaround changes sceCdGetReadPos() to always return 0 on the EE, so that the game will not observe any incremental updates to reading progress. It will then process all data requested at once, when sceCdSync() indicates completion.

The NTSC-J SLPS-20001 release seemed to not be affected, so the patch is not applied for it at this moment.