Added patch for Harvest Moon: A Wonderful Life.
Harvest Moon create alot of threads. When the game gets stuck, all the threads are either suspended or waiting.
What seems to be happening is thread 16 trying to wake up the main thread. However, this might be failing because
the game also has additional checks around WakeupThread, to prevent a thread from waking up another thread
if the thread to be woken up is not in a Wait state. So if the main thread has not slept, the main thread will
enter a state of eternal sleep once it does because thread 16 will be prohibited from waking it up.

Thread 16 will not make further attempts to wake up the main thread as it will also sleep,
regardless of whether the main thread was successfully woken up or not.

Disabling the extra checks seem to make the glitch go away, but I don't know why they added extra code like that.
By design, calling WakeupThread on a thread that is not sleeping will increase the WakeUpCount value of the thread,
which will cause it to inhibit a number of calls to SleepThread() by that amount
(which I think is fair - if the thread has to be woken up, it should be).

The mistake may be made in various places, but seems to only affect the game's initial loading screen.