Increase read delay cycles for Kingdom Hearts II
The initial version of KH2 has some timing issues with disc reads
which leads to a game lockup when entering the gummi ship menu.

A patch applied by the ee_core exists since quite some time, but
probably because of improvements to the rest of the project, the
original delay cycles count is no longer sufficient to prevent
the game from locking up.

Some bisection has shown that increasing the value from 0x100000
to 0x110000 is sufficient. For reference, a value 0x10c000 already
fails to work.

Also note that there are multiple critical transitions the game does.
One is the obvious transition from the overworld map to the gummi
menu. Another one is the transition from the gummi stage back to the
gummi menu. E.g. with a value of 0x10c000 the former one works, but
the latter fails.

All tests were done using SMB mode and the SLUS_210.05 release. It
seems plausible that the other language release suffer from the same
issues, so apply the new value to all of them.