PADEMU: Support PS2 Guitars
Allow using PS3 Guitar Hero Guitars as PS2 Guitar Hero Guitars

Add in a pad_get_model function that allows for getting the model when
querying the model and make, as the Guitar Hero games use this for
identifying Guitar Hero Guitars. Also, force holding down D-pad left as
this is also used by the games for identifying Guitars.